Refactor: resolvet the  dependency about status_code
